add feature comfirm course for admin 9/6/2025

+++ src/main/java/com/elearning/haui/service/FavoriteCourseService.java
@@ -25,6 +25,10 @@ public class FavoriteCourseService {
     }
 
     public FavoriteCourse addFavoriteCourse(FavoriteCourse favoriteCourse) {
+        if(favoriteCourse.getCourse().getApprovalStatus().equals("pending")||favoriteCourse.getCourse().getApprovalStatus().equals("rejected"))
+        {
+            throw new RuntimeException("This course is not approved");
+        }
         CartDetail cartDetail=cartDetailRepository.getCartDetailByCourseId(favoriteCourse.getCourse().getCourseId());
         if(cartDetail!=null){
             cartDetailRepository.delete(cartDetail);
